refined verbosity and logging
